+++ examples/docs/bookstore/entities/CustomerName.ts
@@ -1,0 +1,6 @@
+export const CustomerNameSchema = {
+  '$id': 'urn:bookstore:CustomerName',
+  'maxLength': 200,
+  'minLength': 1,
+  'type': 'string'
+} as const;