+++ examples/docs/bookstore/entities/Money.ts
@@ -1,5 +1,15 @@
+import { AmountSchema } from './Amount.js';
+import { CurrencyCodeSchema } from './CurrencyCode.js';
+
 export const MoneySchema = {
   '$id': 'urn:bookstore:Money',
-  'minimum': 0,
-  'type': 'number'
+  'properties': {
+    'amount': { '$ref': AmountSchema.$id },
+    'currency': { '$ref': CurrencyCodeSchema.$id }
+  },
+  'required': [
+    'amount',
+    'currency'
+  ],
+  'type': 'object'
 } as const;