+++ examples/docs/bookstore/entities/Order.ts
@@ -1,24 +1,22 @@
+import { AddressSchema } from './Address.js';
 import { CustomerIdSchema } from './CustomerId.js';
-import { CurrencyCodeSchema } from './CurrencyCode.js';
 import { Iso8601Schema } from './Iso8601.js';
 import { MoneySchema } from './Money.js';
 import { OrderIdSchema } from './OrderId.js';
+import { OrderLineSchema } from './OrderLine.js';
 
 export const OrderSchema = {
   '$id': 'urn:bookstore:Order',
   'properties': {
-    'currency': {
-      '$ref': CurrencyCodeSchema.$id,
-      'default': 'USD'
-    },
     'customerId': { '$ref': CustomerIdSchema.$id },
     'id': { '$ref': OrderIdSchema.$id },
     'items': {
-      'items': { '$ref': 'urn:bookstore:OrderLine' },
+      'items': { '$ref': OrderLineSchema.$id },
       'minItems': 1,
       'type': 'array'
     },
     'placedAt': { '$ref': Iso8601Schema.$id },
+    'shippingAddress': { '$ref': AddressSchema.$id },
     'total': { '$ref': MoneySchema.$id }
   },
   'required': [
@@ -26,7 +24,8 @@ export const OrderSchema = {
     'customerId',
     'items',
     'total',
-    'placedAt'
+    'placedAt',
+    'shippingAddress'
   ],
   'type': 'object'
 } as const;