+++ examples/docs/bookstore/entities/PersonName.ts
@@ -1,6 +1,0 @@
-export const PersonNameSchema = {
-  '$id': 'urn:bookstore:PersonName',
-  'maxLength': 200,
-  'minLength': 1,
-  'type': 'string'
-} as const;